[#2524] move unused code
