+++ profiler/src/test/java/com/navercorp/pinpoint/profiler/plugin/xml/GetterInjector.java
