+++ profiler/src/test/java/com/navercorp/pinpoint/profiler/plugin/xml/interceptor/SharedAnnotatedInterceptorInjector.java
